+++ R/edstr-package.R
@@ -97,6 +97,7 @@
 #' @importFrom rlang set_names
 #' @importFrom stringi stri_c stri_detect_regex stri_replace_all_fixed
 #' @importFrom stringi stri_replace_all_regex stri_sub stri_trans_general
+#' @importFrom stringr fixed
 #' @importFrom stringr regex
 #' @importFrom stringr str_detect
 #' @importFrom stringr str_ends